Refactor for reuse in print layout
